refactored the widget code and some small improvements
